+++ web_logo_hostname/migrations/8.0.1.0.8/pre-migration.py
@@ -18,37 +18,6 @@
 #
 ##############################################################################
 
-
 def migrate(cr, version):
     if not version:
         return
-
-    cr.execute("""
-        SELECT c.id, c.hostname
-        FROM res_company c
-    """)
-    hostnames = cr.fetchall()
-
-    for hostname in hostnames:
-        if not hostname[1]:
-            continue
-
-        cr.execute("""
-            SELECT h.id, h.hostname
-            FROM res_company h
-            WHERE h.id = %d
-        """ % hostname[0])
-        host = cr.fetchone()
-
-        if host:
-            # Insert old value to new table
-            cr.execute("""
-                INSERT INTO res_company_hostname(hostname, company_id)
-                VALUES ('%s', %d)
-            """ % (hostname[1], host[0]))
-
-    # Drop old column
-    cr.execute("""
-        ALTER TABLE res_company
-        DROP COLUMN hostname
-    """)